Add blank profile image
From: https://pixabay.com/vectors/blank-profile-picture-mystery-man-973460/
Usage: https://pixabay.com/service/license/

---

Simplified Pixabay License

Our license empowers creators and protects our community. We want to keep it as simple as possible. Here is an overview of what Pixabay content can and can't be used for.

What is allowed?

✓	All content on Pixabay can be used for free for commercial and noncommercial use across print and digital, except in the cases mentioned in "What is not allowed".
✓	Attribution is not required. Giving credit to the contributor or Pixabay is not necessary but is always appreciated by our community.
✓	You can make modifications to content from Pixabay.

What is not allowed?

This section only applies to image users and not to the appropriate image authors.

✕	Don't redistribute or sell someone else's Pixabay images or videos on other stock or wallpaper platforms.
✕	Don't sell unaltered copies of an image. e.g. sell an exact copy of a stock photo as a poster, print or on a physical product.
✕	Don't portray identifiable people in a bad light or in a way that is offensive.
✕	Don't use images with identifiable brands to create a misleading association with a product or service.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>