Add per-game locking
This should allow games to progress concurrently of each other.
Previously we held the global controller for all dispatch operations
even though we already hold a pointer to the GameData. Since this
pointer won't change and at worst it'll disappear (with a GC language we
don't care too much -- there's no deconstructor), this should be safe.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>