Fix issue with wrong user display
Allow lookup of user by APIToken, when no other identifier is specified.
This lets us compare the stored APIToken and user account information to
the latest data from the remote. It is unlikely the user ids will be the
same in the event of the user account being invalid. Additionally, API
will now 404 when no user is found. This further lets us validate
whether or not this API token is still valid.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>