Prevent users from arbitrarily entering games
This adds support for controlling which users can view/enter a game
without knowing its join code and only by game identifier. By using the
GamePlayer object, we can track per-user join codes and whether or not
they're added to the game. When they're not, the websocket returns 403
(permission denied).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>