Fix websocket hang caused by Notification queue
This probably fixes Matt's issue as well when he spams peek at the
websocket. Because other players can't keep up (especially after he
disconnects one or more of his players), we quit being able to send
messages and the entire game hangs because it is waiting on a websocket
to drain the Notification channel, which will never get drained as the
player has left. Instead, check that Notifications is non-nil before
writing to it, and allow the WebSocket layer to remove Notifications
when the player disconnects.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>