Update user's game after being admitted
This would let them see the configuration of the game rather than having
a default configuration. However, this will need some server updates in
order to function: being admitted (in the Websocket Controller) doesn't
yet flush to the database, resulting in an out of date view when the
user hits the API again.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>