Always show correct winner in messages
In Rush, the last person to play is the winner. In other games however,
we still need to declare a winner; this winner could be decided after
scoring (in the case of 3 13) -- but scoring and play turn doesn't
necessarily indicate a winner. Load the winner information from the game
state, rather than from the current player.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>