Add server-side rate limiting
This adds a 7 req/sec rate limit on the read side of the connection. Our
assumption is that this is sufficient to prevent abuse of a given
websocket -- but we shall see if it is perhaps too few requests in
practice.

Since each game has its own websocket, it shouldn't be a problem for
users playing multiple games.

Resolves: #109

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>