Add figgy: a game configuration helper
Figgy is a game configuration helper that walks struct tags and uses
them to validate configuration options, assign labels, build
configuration descriptions, and load configuration from JSON or maps.
This will be used in game configuration by:

 1. Automatically loading game configuration options from the server,
 2. Having a single source of truth for things like default values and
    limits,
 3. Allow us to more uniformly parse configuration from a dump on disk.

This should ease the speed at which we add options and game
configuration in the future.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>