[figgy] Allow read-only recursion for validation
When validating a member of a struct, we'd continue, silently ignoring
the field, because we're unable to set it. This is because validating a
config sometimes takes the config itself and not a reference to it. So,
add a read-only member for the recursion helper, to let us know if we
are allowed to continue (without causing a panic later because we
attempt to set a value we can only read).

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>