Enable multiple websocket sessions for the same user
This works towards allowing users to join (and see useful state) from
multiple devices. There's clearly a lot more work to do in all of the
games and admin interfaces to allow the devices to show state in sync
with each other, but otherwise it generally works and should allow for a
little more graceful transitions between devices.

Resolves: #206

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>