Fix selection for self-playing
Due to the change in 0268e7780a69303e11e82d4625de1bfdb4b349d4, we now
always send select events. When the player is clicking around the board
idly, but it isn't their turn, we'd previously fail to update the
current player's selection because it wasn't their turn. In this case,
when a selection event comes in from the current player and it isn't
necessarily their turn, update the incoming player's selection instead
since they're not a spectator.

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>