8Js: Ignore selection events reflected from us
When the local client sends selection events up to the server, they get
reflected back to us, though this can sometimes lag and the user
might've selected a different time on this client in the mean time. As a
result, we should ignore any delayed, reflected selection events and
wait for the server to catch up.

Resolves: #239

Signed-off-by: Alexander Scheel <alexander.m.scheel@gmail.com>