Add test case, fix bug
We were allowed to require more wildcards than cards, erroneously
returning a negative value for `more` instead of failing the match.

This did not affect the actual test case that was added, but I
noticed the discrepancy in the logs. It only affected one existing
test case which was wrong lol.